requirements.txt
fastapi
uvicorn
fastf1
pandas
sqlalchemy
psycopg2-binary
python-dotenv
